+++ boards/twrkv58f220m/freemaster_examples/fmstr_can/board.h
@@ -21,13 +21,13 @@
 
 /*! @brief The UART to use for debug messages. */
 #define BOARD_USE_UART
-#define BOARD_DEBUG_UART_TYPE kSerialPort_Uart
-#define BOARD_DEBUG_UART_BASEADDR UART0
+#define BOARD_DEBUG_UART_TYPE     kSerialPort_Uart
+#define BOARD_DEBUG_UART_BASEADDR (uint32_t) UART0
 #define BOARD_DEBUG_UART_INSTANCE 0U
-#define BOARD_DEBUG_UART_CLKSRC kCLOCK_FastPeriphClk
+#define BOARD_DEBUG_UART_CLKSRC   kCLOCK_FastPeriphClk
 #define BOARD_DEBUG_UART_CLK_FREQ CLOCK_GetFastPeriphClkFreq()
-#define BOARD_UART_IRQ UART0_RX_TX_IRQn
-#define BOARD_UART_IRQ_HANDLER UART0_RX_TX_IRQHandler
+#define BOARD_UART_IRQ            UART0_RX_TX_IRQn
+#define BOARD_UART_IRQ_HANDLER    UART0_RX_TX_IRQHandler
 
 #ifndef BOARD_DEBUG_UART_BAUDRATE
 #define BOARD_DEBUG_UART_BAUDRATE 115200
@@ -40,11 +40,11 @@
 #define BOARD_ENET0_PHY_ADDRESS (0x00U) /* Phy address of enet port 0. */
 
 /*! @brief The bubble level demo information */
-#define BOARD_FXOS8700_ADDR 0x1D
-#define BOARD_ACCEL_ADDR BOARD_FXOS8700_ADDR
-#define BOARD_ACCEL_BAUDRATE 100
-#define BOARD_ACCEL_I2C_BASEADDR I2C1
-#define BOARD_ACCEL_I2C_CLK_SRC I2C1_CLK_SRC
+#define BOARD_FXOS8700_ADDR        0x1D
+#define BOARD_ACCEL_ADDR           BOARD_FXOS8700_ADDR
+#define BOARD_ACCEL_BAUDRATE       100
+#define BOARD_ACCEL_I2C_BASEADDR   I2C1
+#define BOARD_ACCEL_I2C_CLK_SRC    I2C1_CLK_SRC
 #define BOARD_ACCEL_I2C_CLOCK_FREQ CLOCK_GetFreq(I2C1_CLK_SRC)
 
 /*! @brief The dspi instance used for board. */
@@ -60,9 +60,9 @@
 #ifndef BOARD_SW2_GPIO_PIN
 #define BOARD_SW2_GPIO_PIN 4U
 #endif
-#define BOARD_SW2_IRQ PORTA_IRQn
+#define BOARD_SW2_IRQ         PORTA_IRQn
 #define BOARD_SW2_IRQ_HANDLER PORTA_IRQHandler
-#define BOARD_SW2_NAME "SW2"
+#define BOARD_SW2_NAME        "SW2"
 
 #ifndef BOARD_SW3_GPIO
 #define BOARD_SW3_GPIO GPIOE
@@ -73,12 +73,12 @@
 #ifndef BOARD_SW3_GPIO_PIN
 #define BOARD_SW3_GPIO_PIN 4U
 #endif
-#define BOARD_SW3_IRQ PORTE_IRQn
+#define BOARD_SW3_IRQ         PORTE_IRQn
 #define BOARD_SW3_IRQ_HANDLER PORTE_IRQHandler
-#define BOARD_SW3_NAME "SW3"
+#define BOARD_SW3_NAME        "SW3"
 
 /* Board led color mapping */
-#define LOGIC_LED_ON 0U
+#define LOGIC_LED_ON  0U
 #define LOGIC_LED_OFF 1U
 #ifndef BOARD_LED_RED_GPIO
 #define BOARD_LED_RED_GPIO GPIOE
@@ -111,9 +111,9 @@
 
 #define LED_RED_INIT(output)                                           \
     GPIO_PinWrite(BOARD_LED_RED_GPIO, BOARD_LED_RED_GPIO_PIN, output); \
-    BOARD_LED_RED_GPIO->PDDR |= (1U << BOARD_LED_RED_GPIO_PIN)                        /*!< Enable target LED_RED */
-#define LED_RED_ON() GPIO_PortClear(BOARD_LED_RED_GPIO, 1U << BOARD_LED_RED_GPIO_PIN) /*!< Turn on target LED_RED */
-#define LED_RED_OFF() GPIO_PortSet(BOARD_LED_RED_GPIO, 1U << BOARD_LED_RED_GPIO_PIN)  /*!< Turn off target LED_RED */
+    BOARD_LED_RED_GPIO->PDDR |= (1U << BOARD_LED_RED_GPIO_PIN)                         /*!< Enable target LED_RED */
+#define LED_RED_ON()  GPIO_PortClear(BOARD_LED_RED_GPIO, 1U << BOARD_LED_RED_GPIO_PIN) /*!< Turn on target LED_RED */
+#define LED_RED_OFF() GPIO_PortSet(BOARD_LED_RED_GPIO, 1U << BOARD_LED_RED_GPIO_PIN)   /*!< Turn off target LED_RED */
 #define LED_RED_TOGGLE() \
     GPIO_PortToggle(BOARD_LED_RED_GPIO, 1U << BOARD_LED_RED_GPIO_PIN) /*!< Toggle on target LED_RED */
